+++ module/applications/_module.mjs
@@ -1,3 +1,4 @@
 export * as apps from "./apps/_module.mjs";
 export * as sheets from "./sheets/_module.mjs";
 export * as sidebar from "./sidebar/_module.mjs";
+export * as ux from "./ux/_module.mjs";
