+++ module/chat.js
@@ -1,6 +1,5 @@
-import { handleRoll } from "./enrichers/roll.js";
+import { handleRoll } from "./applications/ux/enrichers/roll.mjs";
 import { Helper } from "./helper.js";
-import { MultiAttackRoll } from "./roll/multi-attack-roll.js";
 
 /**
  * Highlight critical success or failure on d20 rolls, or recharge rolls
