+++ module/dice/_module.mjs
@@ -1,0 +1,1 @@
+export { default as Roll4e } from "./Roll.mjs";
